+++ test/cljs/lambdacd/pipeline_test.cljs
@@ -1,6 +1,6 @@
 (ns lambdacd.pipeline-test
   (:require [cljs.test :refer-macros [deftest is testing run-tests]]
-            [lambdacd.testdata :refer [some-build-step some-build-step-id with-name with-type with-output with-children time-start time-after-ten-sec]]
+            [lambdacd.testdata :refer [some-build-step some-build-step-id with-name with-type with-output with-children with-trigger-id with-status time-start time-after-ten-sec]]
             [dommy.core :as dommy]
             [dommy.core :refer-macros [sel sel1]]
             [lambdacd.pipeline :as pipeline]
@@ -21,6 +21,23 @@
       (with-children [some-build-step])
       (with-output "hello from container")))
 
+(def some-build-step-with-custom-type
+  (-> some-build-step
+      (with-type "some-type")))
+
+(def some-manual-trigger
+  (-> some-build-step
+      (with-name "some-manual-trigger")
+      (with-type "manual-trigger")
+      (with-trigger-id "some-trigger-id")
+      (with-status "waiting")))
+
+(def some-inactive-manual-trigger
+  (-> some-build-step
+      (with-name "some-inactive-manual-trigger")
+      (with-type "manual-trigger")
+      (with-status nil)))
+
 (def some-parallel-build-step
   (-> some-container-build-step
       (with-name "some-parallel-step")
@@ -30,6 +47,9 @@
 
 (defn steps [root]
   (sel root :li))
+
+(defn icons [root]
+  (sel root :i))
 
 (defn step-label [step]
   (sel1 step :span))
@@ -78,6 +98,14 @@
           (is (dom/having-class "build-step" (step-label (first (steps div)))))
           (is (dom/having-data "status" "success" (first (steps div))))
           (is (dom/containing-link-to div (route/for-build-and-step-id some-build-number [1 2 3]))))))
+     (testing "rendering of a single build-step with custom display-type"
+      (tu/with-mounted-component
+        [pipeline/build-step some-build-step-with-custom-type some-build-number]
+        (fn [c div]
+          (is (dom/found-in div #"some-step"))
+          (is (dom/having-class "build-step" (step-label (first (steps div)))))
+          (is (dom/having-data "status" "success" (first (steps div))))
+          (is (dom/containing-link-to div (route/for-build-and-step-id some-build-number [1 2 3]))))))
     (testing "rendering of a container build-step"
       (tu/with-mounted-component
         [pipeline/build-step some-container-build-step some-build-number]
@@ -95,7 +123,22 @@
           (is (dom/found-in (first (steps div)) #"some-other-step"))
           (is (dom/having-class "build-step" (step-label (first (steps div)))))
           (is (dom/having-data "status" "success" (first (steps div))))
-          (is (dom/containing-unordered-list (first (steps div)))))))))
+          (is (dom/containing-unordered-list (first (steps div)))))))
+    (testing "rendering of a waiting manual trigger"
+      (tu/with-mounted-component
+        [pipeline/build-step some-manual-trigger some-build-number]
+        (fn [c div]
+          (is (dom/found-in div #"some-manual-trigger"))
+          (is (dom/having-class "build-step" (step-label (first (steps div)))))
+          (is (dom/having-data "status" "waiting" (first (steps div))))
+          (is (dom/having-class "fa-play" (first (icons div)))))))
+    (testing "rendering of an inactive manual trigger"
+      (tu/with-mounted-component
+        [pipeline/build-step some-inactive-manual-trigger some-build-number]
+        (fn [c div]
+          (is (dom/found-in div #"some-inactive-manual-trigger"))
+          (is (dom/having-class "fa-play" (first (icons div))))
+          (is (dom/having-class "pipeline__step__action-button--disabled" (first (icons div)))))))))
 
 (deftest pipeline-test
   (testing "rendering of a complete pipeline"