+++ CHANGELOG.md
@@ -6,6 +6,8 @@ However, as this is still an experimental library, breaking changes may occur wi
 The official release will have a defined and more stable API. If you are already relying on a particular API, please let me know.
 
 # 0.7.2
+* Improvements: 
+  * UI: Trigger symbol is now visible before a manual trigger is reached (#97)
 * Bug fixes:
   * UI did not display the second detail if it had the same label as the first (#98)